# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CheckoutInputColor = typeof CheckoutInputColor[keyof typeof CheckoutInputColor];


export const CheckoutInputColor = {
  White: 'White',
  Black: 'Black',
} as const;

export interface CheckoutInput {
  email: string;
  phone?: string;
  color: CheckoutInputColor;
}

export interface CheckoutData {
  checkoutUrl: string;
}

export interface CheckoutResponse {
  data: CheckoutData;
}