# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: checkout
    description: FitX checkout operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /checkout/create:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Start the FitX checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Invalid checkout information
        "429":
          description: Too many checkout attempts
        "502":
          description: Checkout provider unavailable
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CheckoutInput:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        color:
          type: string
          enum: [White, Black]
      required:
        - email
        - color
    CheckoutData:
      type: object
      properties:
        checkoutUrl:
          type: string
          format: uri
      required:
        - checkoutUrl
    CheckoutResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/CheckoutData"
      required:
        - data