# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutInputColor } from './checkoutInputColor';

export interface CheckoutInput {
  email: string;
  phone?: string;
  color: CheckoutInputColor;
}