# lib/api-zod/src/generated/types/checkoutResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutData } from './checkoutData';

export interface CheckoutResponse {
  data: CheckoutData;
}